переделал меню в файл
