+++ MySite/Blog/static/blog/css/modal.css
@@ -13,7 +13,7 @@
 }
 
 .modal-content {
-    margin: 20px auto 0;
+    margin: 70px auto 0;
     display: block;
     max-width: 80%;
     max-height: 80%;
