+++ MySite/Blog/templates/blog/edit_posts_form.html
@@ -5,25 +5,32 @@
 
 
 {% if messages %}
-    <ul class="messages">
-        {% for message in messages %}
-            <li {% if message.tags %} class="{{ message.tags }}"{% endif %}>
-                {{ message }}
-            </li>
-        {% endfor %}
-    </ul>
+<ul class="messages">
+    {% for message in messages %}
+    <li {% if message.tags %} class="{{ message.tags }}" {% endif %}>
+        {{ message }}
+    </li>
+    {% endfor %}
+</ul>
 {% endif %}
 
-    <div class="content-section">
-        <form method="POST" enctype="multipart/form-data">
-            {% csrf_token %}
-            <fieldset class="form-group"  >
-                <legend class="border-bottom mb-4">Редактирование </legend>
-                {{ form|crispy }}
-            </fieldset>
-            <div class="form-group">
-                <button class="btn btn-outline-info" type="submit">Сохранить</button>
-            </div>
-        </form>
-    </div>
+<div class="content-section">
+    <form method="POST" enctype="multipart/form-data">
+        {% csrf_token %}
+        <fieldset class="form-group">
+            <legend class="border-bottom mb-4">Редактирование</legend>
+            {{ form|crispy }}
+        </fieldset>
+
+        <div class="row">
+            {% for image in post_images %}
+            <li class="image-post-detail"> <img src="{{ image.image.url }}" alt=""></img></li>
+            {% endfor %}
+        </div>
+
+        <div class="form-group">
+            <button class="btn btn-outline-info" type="submit">Сохранить</button>
+        </div>
+    </form>
+</div>
 {% endblock content %}
