+++ MySite/MySite/urls.py
@@ -3,7 +3,7 @@ from django.contrib.auth import views as auth_views
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
-from users import views as user_views
+
 
 urlpatterns = [
     path('admin/', admin.site.urls),
