+++ MySite/users/templates/users/custom_clearable_file_input.html
@@ -1,0 +1,9 @@
+<div class="custom-clearable-file-input">
+  <label class="custom-clearable-file-input-label" for="{{ widget.attrs.id }}">
+    <img class="custom-clearable-file-input-preview" src="{{ widget.value.url }}" alt="Фото">
+  </label>
+  <input type="{{ widget.type }}" name="{{ widget.name }}" class="custom-clearable-file-input-file" id="{{ widget.attrs.id }}">
+  <span class="custom-clearable-file-input-clear">
+    <label for="{{ widget.attrs.id }}" class="custom-clearable-file-input-clear-label">Изменить</label>
+  </span>
+</div>
