+++ MySite/users/templates/users/profile.html
@@ -3,6 +3,7 @@
 {% extends "blog/base.html" %}
 {% load crispy_forms_tags %}
 {% load custom_tags %}
+
 {% block content %}
 
 {% if messages %}
@@ -16,14 +17,44 @@
     {% endfor %}
 </ul>
 {% endif %}
+<!-- Ссылка на уведомления с восклицательным знаком -->
+<a href="{% url 'show_notifications' %}" class="btn btn-info">
+    {% if profile.notifications %}
+    Мои уведомления <span class="badge badge-danger">!</span>
+    {% else %}
+    Мои уведомления
+    {% endif %}
+</a>
 
 <table border=0 cellpadding="0" cellspacing="0">
-<h2 class="border-bottom mb-4">Ваш профиль</h2>
+    <h2 class="border-bottom mb-4">Ваш профиль</h2>
     <tr>
         <th>
             <div class="content-section">
+
                 <div class="media">
-                    <!-- Ваш код для отображения профиля -->
+                    <!--                    <img class="rounded-circle account-img" src="{{ user.profile.image.url }}">-->
+                    <div class="media-body">
+                        <h3 class="account-heading">Логин: {{ user.username }}</h3>
+                        <!--                        <h3 class="account-heading">Имя: {{ user.first_name }} {{ user.last_name }}</h3>-->
+
+                        <p class="text-secondary">Зарегистрирован: {{ user.date_joined }}</p>
+                        <p class="text-secondary">Ваш Email: {{ user.email }}</p>
+                        <p class="text-secondary">Отзывы: {{ profile.review }}</p>
+                        <p class="text-secondary">Рейтинг: {% if profile.rating == 0 %}
+                            Пока нет рейтинга
+                            {% else %}
+                            {{ profile.rating }}
+                            {% endif%}
+                        </p>
+
+                        <p class="text-secondary">Статус: {% if profile.is_seller %}
+                            Продавец
+                            {% else %}
+                            Покупатель
+                            {% endif%}
+                        </p>
+                    </div>
                 </div>
 
                 <!-- Форма обновления профиля -->
@@ -39,14 +70,7 @@
                     </div>
                 </form>
 
-                <!-- Ссылка на уведомления с восклицательным знаком -->
-                <a href="{% url 'show_notifications' %}" class="btn btn-info">
-                    {% if profile.notifications %}
-                        Мои уведомления <span class="badge badge-danger">!</span>
-                    {% else %}
-                        Мои уведомления
-                    {% endif %}
-                </a>
+
             </div>
         </th>
     </tr>
