+++ MySite/users/views.py
@@ -1,21 +1,12 @@
+from Blog.menu import menu
 from django.contrib import messages
 from django.contrib.auth import login
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from django.shortcuts import redirect, render, get_object_or_404
-
-from Blog.utils import menu
 from users.choices.city_choices import CITY_CHOICES
 from users.forms import UserRegisterForm, UserUpdateForm, ProfileUpdateForm
 from users.models import Profile
-
-
-# @receiver(post_save, sender=User)
-# def create_user_profile(sender, instance, created, **kwargs):
-#     if created:
-#         Profile.objects.create(user=instance)
 
 
 def register(request):
